+++ DesignToolbox/DesignToolbox/Pages/Tokens/Color/NamedColor/NamedColor+Border.swift
@@ -17,18 +17,31 @@ extension NamedColor {
 
     enum Border: String, CaseIterable {
         case borderBrandPrimary
+        case borderBrandSecondary
+        case borderBrandTertiary
         case borderDefault
         case borderEmphasized
         case borderMuted
         case borderFocus
         case borderFocusInset
         case borderOnBrandPrimary
+        case borderOnBrandSecondary
+        case borderOnBrandTertiary
         case borderMinimal
+        case borderStatusPositive
+        case borderStatusInfo
+        case borderStatusWarning
+        case borderStatusNegative
+        case borderStatusAccent
 
         func token(from theme: OUDSTheme) -> MultipleColorSemanticTokens {
             switch self {
             case .borderBrandPrimary:
                 theme.colors.borderBrandPrimary
+            case .borderBrandSecondary:
+                theme.colors.borderBrandSecondary
+            case .borderBrandTertiary:
+                theme.colors.borderBrandTertiary
             case .borderDefault:
                 theme.colors.borderDefault
             case .borderEmphasized:
@@ -41,8 +54,22 @@ extension NamedColor {
                 theme.colors.borderFocusInset
             case .borderOnBrandPrimary:
                 theme.colors.borderOnBrandPrimary
+            case .borderOnBrandSecondary:
+                theme.colors.borderOnBrandSecondary
+            case .borderOnBrandTertiary:
+                theme.colors.borderOnBrandTertiary
             case .borderMinimal:
                 theme.colors.borderMinimal
+            case .borderStatusPositive:
+                theme.colors.borderStatusPositive
+            case .borderStatusInfo:
+                theme.colors.borderStatusInfo
+            case .borderStatusWarning:
+                theme.colors.borderStatusWarning
+            case .borderStatusNegative:
+                theme.colors.borderStatusNegative
+            case .borderStatusAccent:
+                theme.colors.borderStatusAccent
             }
         }
     }