+++ DesignToolbox/DesignToolbox/Pages/Tokens/Color/NamedColor/NamedColor+Content.swift
@@ -13,10 +13,13 @@
 
 import OUDSSwiftUI
 
+// swiftlint:disable function_body_length
 extension NamedColor {
 
     enum Content: String, CaseIterable {
         case contentBrandPrimary
+        case contentBrandSecondary
+        case contentBrandTertiary
         case contentDefault
         case contentDisabled
         case contentMuted
@@ -29,17 +32,33 @@ extension NamedColor {
         case contentOnActionPressed
         case contentOnActionSelected
         case contentOnBrandPrimary
+        case contentOnBrandSecondary
+        case contentOnBrandTertiary
         case contentInverse
+        case contentOnStatusPositiveMuted
+        case contentOnStatusPositiveEmphasized
+        case contentOnStatusNegativeMuted
         case contentOnStatusNegativeEmphasized
+        case contentOnStatusWarningMuted
+        case contentOnStatusWarningEmphasized
+        case contentOnStatusAccentMuted
+        case contentOnStatusAccentEmphasized
+        case contentOnStatusInfoMuted
+        case contentOnStatusInfoEmphasized
         case contentStatusInfo
         case contentStatusNegative
         case contentStatusPositive
         case contentStatusWarning
+        case contentStatusAccent
 
         func token(from theme: OUDSTheme) -> MultipleColorSemanticTokens {
             switch self {
             case .contentBrandPrimary:
                 theme.colors.contentBrandPrimary
+            case .contentBrandSecondary:
+                theme.colors.contentBrandSecondary
+            case .contentBrandTertiary:
+                theme.colors.contentBrandTertiary
             case .contentDefault:
                 theme.colors.contentDefault
             case .contentDisabled:
@@ -64,10 +83,32 @@ extension NamedColor {
                 theme.colors.contentOnActionSelected
             case .contentOnBrandPrimary:
                 theme.colors.contentOnBrandPrimary
+            case .contentOnBrandSecondary:
+                theme.colors.contentOnBrandSecondary
+            case .contentOnBrandTertiary:
+                theme.colors.contentOnBrandTertiary
             case .contentInverse:
                 theme.colors.contentInverse
+            case .contentOnStatusPositiveMuted:
+                theme.colors.contentOnStatusPositiveMuted
+            case .contentOnStatusPositiveEmphasized:
+                theme.colors.contentOnStatusPositiveEmphasized
+            case .contentOnStatusNegativeMuted:
+                theme.colors.contentOnStatusNegativeMuted
             case .contentOnStatusNegativeEmphasized:
                 theme.colors.contentOnStatusNegativeEmphasized
+            case .contentOnStatusWarningMuted:
+                theme.colors.contentOnStatusWarningMuted
+            case .contentOnStatusWarningEmphasized:
+                theme.colors.contentOnStatusWarningEmphasized
+            case .contentOnStatusAccentMuted:
+                theme.colors.contentOnStatusAccentMuted
+            case .contentOnStatusAccentEmphasized:
+                theme.colors.contentOnStatusAccentEmphasized
+            case .contentOnStatusInfoMuted:
+                theme.colors.contentOnStatusInfoMuted
+            case .contentOnStatusInfoEmphasized:
+                theme.colors.contentOnStatusInfoEmphasized
             case .contentStatusInfo:
                 theme.colors.contentStatusInfo
             case .contentStatusNegative:
@@ -76,7 +117,11 @@ extension NamedColor {
                 theme.colors.contentStatusPositive
             case .contentStatusWarning:
                 theme.colors.contentStatusWarning
+            case .contentStatusAccent:
+                theme.colors.contentStatusAccent
             }
         }
     }
 }
+
+// swiftlint:enable function_body_length