+++ DesignToolbox/DesignToolbox/Pages/Tokens/Dimension/Size/SizeTokenPage.swift
@@ -26,6 +26,7 @@ struct SizeTokenPage: View {
                 DesignToolboxCode(code: "theme.iconWithHeadingXLargeSm.dimension(for: horizontalSizeClass ?? .regular)",
                                   titleText: "app_tokens_common_viewCodeExample_label")
             }
+
             Section {
                 VStack(alignment: .center, spacing: theme.spaces.fixedNone) {
                     ForEach(NamedSize.IconDecorative.allCases, id: \.rawValue) { namedSize in
@@ -44,6 +45,26 @@ struct SizeTokenPage: View {
                 }
             } header: {
                 sectionHeader("app_tokens_dimension_size_iconWith_label")
+            }
+
+            Section {
+                VStack(alignment: .leading, spacing: theme.spaces.fixedMedium) {
+                    ForEach(NamedSize.MaxWidth.allCases, id: \.rawValue) { namedSize in
+                        IllustrationMaxWidth(namedSize: namedSize)
+                    }
+                }
+            } header: {
+                sectionHeader("app_tokens_dimension_size_maxWidth_label")
+            }
+
+            Section {
+                VStack(alignment: .leading, spacing: theme.spaces.fixedMedium) {
+                    ForEach(NamedSize.MinWidth.allCases, id: \.rawValue) { namedSize in
+                        IllustrationMinWidth(namedSize: namedSize)
+                    }
+                }
+            } header: {
+                sectionHeader("app_tokens_dimension_size_minWidth_label")
             }
         }
         .padding(.horizontal, theme.spaces.fixedMedium)